More progress primarily in warfare pdus
